Tweak clang-format
The BTF / SCC code is being migrated out of Google source control into
AOSP. This comes with the opportunity to migrate the style to be
closer libabigail's.

In the first instance it makes sense to just reformat it with
clang-format. I've updated this as follows.

AlignConsecutiveDeclarations: false - as this is closer on average to
libgabigail code and looks nicer for the new code anyway

AllowShort*: (true) - these in theory should make things closer to
libabigail style but in practice don't appear make any difference

Cpp11BracedListStyle: true - this seems to improve some initialiser
syntax

BinPackArguments: false - we already turn this off for parameters

SpaceAfterCStyleCast: true - seems to be the libabigail style

	* .clang-format: Various tweaks for new BTF code.

Signed-off-by: Giuliano Procida <gprocida@google.com>