abg-reader: eliminate unreachable-code-loop-increment warnings
This commit tidies up some debris from the change to use
xmlFirstElementChild and xmlNextElementSibling. A couple of while
loops had bodies which would be executed exactly once searching for
the next element with a given tag. These trigger warnings with recent
Clang++, at least.

This commit replaces the loops with their bodies and eliminates
a little redundant conditional logic.

There is no change in behaviour.

	* src/abg-reader.cc (read_translation_unit_from_input): In the
	branch where the context object holds a non-null node, just
	check that it an abi-instr; remove the final null check on
	node as it is guaranteed to be non-null.
	(read_elf_needed_from_input): Save the context object's node
	in node immediately; in the branch where the context object
	holds a non-null node, just check that it an elf-needed;
	remove the final null check on node as it is guaranteed to be
	non-null.

Signed-off-by: Giuliano Procida <gprocida@google.com>