XML writer: fix write_{class,union}_decl state management asymmetry
This may not be the best solution. I think the code change it's
partially undoing was possibly a bad idea.

It also runs the risk of multiple declarations resolving to the same
other one and the latter being emitted multiple times.

This change may have been obsoleted by the change "ensure type ids
correspond to written types".

Signed-off-by: Giuliano Procida <gprocida@google.com>