XML writer: type ids, lies and useless code
Also avoid constructing strings from interned strings.

The code removed seems to be a useless hangover. I wonder if this
change might be rejected or some have assertions added to check the
invariants.

All the test churn appears to be caused by a different ordering of
type id allocation. This should be fixable with a small change, one
which can be reversed once this is proven to be the exact cause.

Signed-off-by: Giuliano Procida <gprocida@google.com>