XML writer: no longer pass type ids to write_ functions
Several of the write_ functions exist in two forms: the first which
wraps the second and the second which takes an extra 'id' argument.
This functionality was first introduced in commit be14ac2ab24d "Fix
hashing of member types" to allow type ids to be specified for
serialisation.

However, the code no longer seems to serve any purpose as no such
overriding ever actually happens. Once it is removed, some calls to
get_id_for_type can be moved back in line and some local variables can
be eliminated.

Note that this change leaves an early type id allocation in
write_member_type. This will be removed in the next change.

No tests are affected and there is no change in behaviour.

	* src/abg-writer.cc (write_class_decl_opening_tag): Remove
	id argument. Move call to get_id_for_type to where needed.
	This avoids create a temporary id string.
	(write_union_decl_opening_tag): Likewise.
	(write_naming_typedef): Avoid creating a temporary id string.
	(write_decl_in_scope): Remove empty string id argument to
	write_class_decl_opening_tag and
	write_union_decl_opening_tag.
	(write_qualified_type_def): Remove id argument and type id
	local variable. Move call to get_id_for_type where needed.
	This avoid creating a temporary string. Drop redundant
	overload.
	(write_pointer_type_def): Likewise.
	(write_reference_type_def): Likewise.
	(write_array_type_def): Likewise.
	(write_enum_type_decl): Likewise.
	(write_typedef_decl): Likewise. Remove another local variable.
	(write_function_type): Move get_id_for_type to where needed.
	(write_class_decl): Remove id argument and no longer pass this
	on to write_class_decl_opening_tag. Drop redundant overload.
	(write_union_decl): Remove id argument and no longer pass this
	on to write_union_decl_opening_tag. Drop redundant overload.
	(write_member_type): No longer pass early-allocated type id to
	callees.
	(write_type_tparameter): Adjust some whitespace.

Reviewed-by: Matthias Maennich <maennich@google.com>
Signed-off-by: Giuliano Procida <gprocida@google.com>