IR: clean up look_through_decl_only
One method was private and had subtly different semantics but was
fortunately was not used except to implement the other two.

Inlined this into a second method (eliminating a conditional) and used
this to implement the third.

It's conceivable the performance of new code is either better or worse
than the old. I've not stared at it for long enough or tested it.

This change needs more scrutiny. It looks like I also removed some
unused functions.

Signed-off-by: Giuliano Procida <gprocida@google.com>