IR: undo naming typedef damage
This is quick attempt at undoing some of the recent naming typedef
changes, namely the bit where the code turns anonymous types into
named types.

I thought it might be needed in this series, but it probably is not.

Note that that change rendered some other logic (around the handling
of anonymous types) dead, but fortunately the dead logic wasn't
cleaned up order there would be more to revert. This raises an
interesting point: if abitidy can undo the damage, what difference
does that logic make?

Signed-off-by: Giuliano Procida <gprocida@google.com>